Refactor project calendar to shared base
