+++ frontend/src/dashboard/project/components/Shared/calendar/constants.ts
@@ -1,0 +1,62 @@
+import type { CategoryOption, UnitOption } from "./types";
+
+export const CATEGORY_OPTIONS: readonly CategoryOption[] = [
+  "AUDIO-VISUAL",
+  "CLIENT-SERVICES-VIP",
+  "CONTINGENCY-MISC",
+  "DECOR",
+  "DESIGN",
+  "FABRICATION",
+  "FOOD-BEVERAGE",
+  "GRAPHICS",
+  "INSTALLATION-MATERIALS",
+  "LABOR",
+  "LIGHTING",
+  "MERCH-SWAG",
+  "PARKING-FUEL-TOLLS",
+  "PERMITS-INSURANCE",
+  "PRODUCTION-MGMT",
+  "RENTALS",
+  "STORAGE",
+  "TECH-INTERACTIVES",
+  "TRAVEL",
+  "TRUCKING",
+  "VENUE-LOCATION-FEES",
+  "WAREHOUSE",
+];
+
+export const UNIT_OPTIONS: readonly UnitOption[] = [
+  "Each",
+  "Hrs",
+  "Days",
+  "EA",
+  "PCS",
+  "Box",
+  "LF",
+  "SQFT",
+  "KG",
+];
+
+export const DOT_SIZE = 10;
+export const DOT_STROKE = 2;
+export const DOT_MAX_VISIBLE = 4;
+export const DOT_OVERLAP_PX = 3;
+
+export const MOBILE_QUERY = "(max-width: 640px)";
+export const POPPER_GAP = 12;
+export const FOCUSABLE_SELECTOR =
+  'a[href], area[href], input:not([disabled]), select:not([disabled]), textarea:not([disabled]), button:not([disabled]), [tabindex]:not([tabindex="-1"])';
+export const INTERACTIVE_SAFE_PAD = 12;
+
+export const WRAPPER_INTERACTIVE_SELECTOR = [
+  "button",
+  "a",
+  "input",
+  "select",
+  "textarea",
+  "[role=\"button\"]",
+  "[role=\"link\"]",
+  "[data-stop-card-nav]",
+  ".calendar-day",
+  ".calendar-day-wrapper",
+].join(", ");
