+++ frontend/src/dashboard/project/components/Shared/calendar/utils.ts
@@ -1,0 +1,34 @@
+export function safeParse(dateStr?: string | null): Date | null {
+  if (!dateStr) return null;
+  if (/^\d{4}-\d{2}-\d{2}$/.test(dateStr)) {
+    const [y, m, d] = dateStr.split("-").map((p) => parseInt(p, 10));
+    return new Date(y, m - 1, d);
+  }
+  const parsed = new Date(dateStr);
+  if (!Number.isNaN(parsed.getTime())) {
+    return new Date(parsed.getFullYear(), parsed.getMonth(), parsed.getDate());
+  }
+  return null;
+}
+
+export function getDateKey(date?: Date | null): string | null {
+  if (!date) return null;
+  const y = date.getFullYear();
+  const m = String(date.getMonth() + 1).padStart(2, "0");
+  const d = String(date.getDate()).padStart(2, "0");
+  return `${y}-${m}-${d}`;
+}
+
+export function formatDateLabel(date: Date): string {
+  return date.toLocaleDateString(undefined, {
+    weekday: "long",
+    month: "long",
+    day: "numeric",
+    year: "numeric",
+  });
+}
+
+export function isElementWithin(node: Node | null, container?: HTMLElement | null): boolean {
+  if (!node || !container) return false;
+  return container === node || container.contains(node as Node);
+}
